Improve page titles and descriptions
On search results, articles will now show the value of the newly created 'seo_title' fromtmatter key. The default title is "Emmanuel Odongo".

In addition the first paragraph of the 'abstract' frontmatter key will be used as the description. The default description is "Emmanuel Odongo is an aspiring web developer who writes about web design and front-end develpoment."