Add title parser to article template
The TITLE in the `middleman article TITLE` command is typically a hyphen separated string. The parser goes over the TITLE, replacing hyphens with spaces and applying a basic title case rule set.

The result of this parsing is set to the frontmatter 'title' and 'seo_title' variables.

Additionaly, a non-breaking-space is added between the final 2 words of parsed title if it meets a certain length criterion. Only the 'title' frontmatter variable an be set to this parsed version.