Fix image links
The asset_hash toggle in the config file adds a few random characters to the end of the file names. To get the proper links to image files, the image_path method should be used instead of a uri.

NOTE: Image links in manifest.json and browserconfig.xml have not been fixed. Both files are in the favicon folder