refactor: use postcss for fluid typography and spacing
With the migration from SvelteKit to Astro, sass was dropped as a
dependency, meaning the `css-lock` mixin could no longer be used. The
native CSS `clamp` function was used in it's place, but it's API is
admittedly clunky.

Fortunately, an alternative has presented itself, in the form of a
PostCSS plugin: postcss-utopia.